telegram-bot/requirements.txt
aiogram
python-dotenv
requests

